move array functions to a submodule
